+++ wordcab_transcribe/main.py
@@ -24,7 +24,8 @@ from fastapi import Depends, FastAPI
 from fastapi import status as http_status
 from fastapi.responses import HTMLResponse
 
-from wordcab_transcribe.config import lifespan, settings
+from wordcab_transcribe.config import settings
+from wordcab_transcribe.dependencies import lifespan
 from wordcab_transcribe.logging import LoggingMiddleware
 from wordcab_transcribe.router.authentication import get_current_user
 from wordcab_transcribe.router.v1.endpoints import (
