+++ wordcab_transcribe/router/v1/youtube_endpoint.py
@@ -27,7 +27,7 @@ from fastapi import APIRouter, BackgroundTasks, HTTPException
 from fastapi import status as http_status
 from loguru import logger
 
-from wordcab_transcribe.config import asr, download_limit
+from wordcab_transcribe.dependencies import asr, download_limit
 from wordcab_transcribe.models import BaseRequest, YouTubeResponse
 from wordcab_transcribe.utils import delete_file, download_audio_file
